requirements.txt
pyspark>=3.5.0,<4.0.0
pyarrow
grpcio
grpcio-status
pandas
python-dotenv
pydantic
boto3